refactor task entity to store date without time
