+++ app/src/main/java/com/giraffe/tudeeapp/data/mapper/TaskMapper.kt
@@ -2,7 +2,8 @@ package com.giraffe.tudeeapp.data.mapper
 
 import com.giraffe.tudeeapp.data.model.TaskEntity
 import com.giraffe.tudeeapp.domain.model.task.Task
-import kotlinx.datetime.LocalDateTime
+import kotlinx.datetime.LocalDate
+import kotlinx.datetime.atTime
 
 
 fun Task.toEntity(): TaskEntity {
@@ -13,9 +14,9 @@ fun Task.toEntity(): TaskEntity {
         taskPriority = this.taskPriority,
         status = this.status,
         categoryId = this.categoryId,
-        dueDate = this.dueDate.toString(),
-        createdAt = this.createdAt.toString(),
-        updatedAt = this.updatedAt.toString()
+        dueDate = this.dueDate.date.atTime(0, 0).toString(),
+        createdAt = this.createdAt.date.atTime(0, 0).toString(),
+        updatedAt = this.updatedAt.date.atTime(0, 0).toString()
     )
 }
 
@@ -27,8 +28,9 @@ fun TaskEntity.toTask(): Task {
         taskPriority = this.taskPriority,
         status = this.status,
         categoryId = this.categoryId,
-        dueDate = LocalDateTime.parse(this.dueDate),
-        createdAt = LocalDateTime.parse(this.createdAt),
-        updatedAt = LocalDateTime.parse(this.updatedAt)
+        dueDate = LocalDate.parse(this.dueDate).atTime(0, 0),
+        createdAt = LocalDate.parse(this.createdAt).atTime(0, 0),
+        updatedAt = LocalDate.parse(this.updatedAt).atTime(0, 0)
     )
 }
+
