+++ app/src/main/java/com/giraffe/tudeeapp/presentation/uimodel/TaskUi.kt
@@ -1,6 +1,5 @@
-package com.giraffe.tudeeapp.presentation.home.taskdetails
+package com.giraffe.tudeeapp.presentation.uimodel
 
-import com.giraffe.tudeeapp.design_system.component.PriorityType
 import com.giraffe.tudeeapp.domain.model.Category
 import com.giraffe.tudeeapp.domain.model.task.Task
 import com.giraffe.tudeeapp.domain.model.task.TaskPriority
@@ -11,7 +10,7 @@ data class TaskUi (
     val id: Long,
     val title: String,
     val description: String,
-    val priorityType: PriorityType,
+    val priorityType: TaskPriority,
     val status: TaskStatus,
     val category: Category,
     val dueDate: LocalDateTime,
@@ -23,16 +22,10 @@ fun Task.toTaskUi(category: Category) = TaskUi(
     id = id,
     title = title,
     description = description,
-    priorityType = taskPriority.toPriorityType(),
+    priorityType = taskPriority,
     status = status,
     category = category,
     dueDate = dueDate,
     createdAt = createdAt,
     updatedAt = updatedAt
 )
-
-fun TaskPriority.toPriorityType() = when(this) {
-    TaskPriority.LOW -> PriorityType.LOW
-    TaskPriority.MEDIUM -> PriorityType.MEDIUM
-    TaskPriority.HIGH -> PriorityType.HIGH
-}
